# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CaReMind API
  version: 1.0.0
  description: REST API for private vehicle, maintenance and operating-cost management.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.car-remind.gr/api
    description: Production
...
paths:
  /login:
    post:
      operationId: login
      tags: [Authentication]
      summary: Sign in and create a refresh session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, description: Username or email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /refresh:
    post:
      operationId: refreshSession
      tags: [Authentication]
      summary: Create a new access token from the HttpOnly refresh cookie
      security: []
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /logout:
    post:
      operationId: logout
      tags: [Authentication]
      summary: Revoke the current refresh session
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /register:
    post:
      operationId: register
      tags: [Authentication]
      summary: Register an account and send an email verification code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Registration' }
      responses:
        '200': { description: Registration accepted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Username or email already exists }
  /verify-email:
    post:
      operationId: verifyEmail
      tags: [Authentication]
      summary: Verify a registration email code
      security: []
      requestBody: { $ref: '#/components/requestBodies/EmailCode' }
      responses:
        '200': { description: Email verified }
        '400': { $ref: '#/components/responses/BadRequest' }
  /resend-verification:
    post:
      operationId: resendVerification
      tags: [Authentication]
      summary: Request a replacement verification code
      security: []
      requestBody: { $ref: '#/components/requestBodies/EmailOnly' }
      responses:
        '200': { description: Generic delivery response }
        '400': { $ref: '#/components/responses/BadRequest' }
  /forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [Authentication]
      summary: Request a password-reset code
      security: []
      requestBody: { $ref: '#/components/requestBodies/EmailOnly' }
      responses:
        '200': { description: Generic delivery response }
        '400': { $ref: '#/components/responses/BadRequest' }
  /verify-reset-code:
    post:
      operationId: verifyPasswordResetCode
      tags: [Authentication]
      summary: Exchange a reset code for a short-lived reset token
      security: []
      requestBody: { $ref: '#/components/requestBodies/EmailCode' }
      responses:
        '200': { description: Reset token created }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /reset-password:
    post:
      operationId: resetPassword
      tags: [Authentication]
      summary: Set a new password and revoke refresh sessions
      security: []
      responses:
        '200': { description: Password updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
...
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  requestBodies:
    EmailOnly:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [email]
            properties: { email: { type: string, format: email } }
    EmailCode:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [email, code]
            properties:
              email: { type: string, format: email }
              code: { type: string, pattern: '^[0-9]{6}$' }
...
  schemas:
    Error:
      type: object
      properties: { error: { type: string }, code: { type: string } }
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: string, enum: [user, admin, guest] }
        companyId: { type: [integer, 'null'] }
        companyName: { type: [string, 'null'] }
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken: { type: string }
        user: { $ref: '#/components/schemas/User' }
    Registration:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 50 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        fullName: { type: string, maxLength: 100 }
        phone: { type: string, maxLength: 50 }
        companyName: { type: string, maxLength: 100 }
        account_type: { type: string, enum: [individual, business] }
...